+++ Core/src/Rendering/SparseVoxelOctree.h
@@ -1,12 +1,16 @@
 #ifndef SPARSE_VOXEL_OCTREE_H
 #define SPARSE_VOXEL_OCTREE_H
 
+#include <GLFW/glfw3.h>
+
 #include "Containers/Array.h"
 #include <Defines.h>
 #include <stddef.h>
 
 #define VOXEL_CHUNK_SIZE 4
 
+// octree element: [octree index of first child][child occupancy mask]
+// has to be 32 bit to
 typedef Array(u32) SparseVoxelOctree;
 
 typedef struct
@@ -15,6 +19,6 @@ typedef struct
 } VoxelChunk;
 
 int CalculateOctreeSize(int denseDataSize);
-void PopulateVoxelChunk(SparseVoxelOctree* result, u32 denseData[]);
+void PopulateVoxelChunk(SparseVoxelOctree* result, u8 denseData[]);
 
 #endif // SPARSE_VOXEL_OCTREE_H
